changed the datatypes

+++ src/main/java/DefaultPriceAndDistances.java
@@ -1,7 +1,7 @@
 public interface DefaultPriceAndDistances {
     int flatPrice=30;
-    int flatPriceDistance =2;
+    float flatPriceDistance =2;
     int firstBreakUpPrice=20;
-    int firstBreakUpDistance=10;
+    float firstBreakUpDistance=10;
     int LastBreakUpPrice =30;
 }
